src/requirements.txt
pip>=21.2.4
google-cloud-storage
boto3